+++ package.json
@@ -28,6 +28,7 @@
     "node-sass": "^6.0.1",
     "nuxt": "^2.15.8",
     "sass-loader": "^10.2.1",
+    "simple-vue-validator": "^0.16.0",
     "sweetalert2": "^11.4.14",
     "vue": "^2.6.14",
     "vue-chartjs": "^4.1.0",
